fwmanager: Extract BootControl into its own module
Move the BootControl trait and HalBootControl out of lib.rs into
boot_control.rs, so that public lib.rs API stays flat
(fwmanager_api::BootControl, ::HalBootControl). Pure move, no code
change; the same BootControl host tests pass.

(cherry picked from commit 68665407fda86b98657e645bc8dda2e96f85ae0d)
